G2-440 Reorganized original manual command converter
