+++ gb-command/src/test/java/com/g2forge/gearbox/command/v2/converter/manual/v1/TestMethodToCommandInvocationTransformer.java
@@ -1,10 +1,15 @@
-package com.g2forge.gearbox.command.v2.outdated;
+package com.g2forge.gearbox.command.v2.converter.manual.v1;
 
 import org.junit.Test;
 
 import com.g2forge.alexandria.java.core.helpers.HCollection;
 import com.g2forge.alexandria.test.HAssert;
 import com.g2forge.gearbox.command.v2.converter.IMethodArgument;
+import com.g2forge.gearbox.command.v2.converter.manual.v1.ArgumentConsumer;
+import com.g2forge.gearbox.command.v2.converter.manual.v1.IArgumentConsumer;
+import com.g2forge.gearbox.command.v2.converter.manual.v1.IMethodConsumer;
+import com.g2forge.gearbox.command.v2.converter.manual.v1.MethodConsumer;
+import com.g2forge.gearbox.command.v2.converter.manual.v1.MethodToCommandInvocationTransformer;
 import com.g2forge.gearbox.command.v2.process.IProcess;
 import com.g2forge.gearbox.command.v2.proxy.method.ITestCommandInterface;
 import com.g2forge.gearbox.command.v2.proxy.method.MethodInvocation;
